hunts show cmd finished, updates will come later. Major code refactor needed

+++ src/hunts/parse.rs
@@ -10,43 +10,43 @@ use serde::{de::Error, Serialize, Deserialize, Deserializer};
 pub struct HuntInfo {
 
   #[serde(rename = "Balance", deserialize_with="de_from_str")]
-  pub balance: i64,
+  pub balance: f64,
 
   #[serde(rename = "Damage", deserialize_with="de_from_str")]
-  pub damage: i64,
+  pub damage: f64,
 
   #[serde(rename = "Damage/h", deserialize_with="de_from_str")]
-  pub damage_h: i64,
+  pub damage_h: f64,
 
   #[serde(rename = "Healing", deserialize_with="de_from_str")]
-  pub healing: i64,
+  pub healing: f64,
 
   #[serde(rename = "Healing/h", deserialize_with="de_from_str")]
-  pub healing_h: i64,
+  pub healing_h: f64,
 
   #[serde(rename = "Killed Monsters")]
   pub killed_monsters: Vec<CountedThing>,
 
   #[serde(rename = "Loot", deserialize_with="de_from_str")]
-  pub loot: i64,
+  pub loot: f64,
 
   #[serde(rename = "Looted Items")]
   pub looted_items: Vec<CountedThing>,
 
   #[serde(rename = "Raw XP Gain", deserialize_with="de_from_str")]
-  pub raw_xp: i64,
+  pub raw_xp: f64,
 
   #[serde(rename = "Raw XP/h", deserialize_with="de_from_str")]
-  pub raw_xp_h: i64,
+  pub raw_xp_h: f64,
 
   #[serde(rename = "Supplies", deserialize_with="de_from_str")]
-  pub supplies: i64,
+  pub supplies: f64,
 
   #[serde(rename = "XP Gain", deserialize_with="de_from_str")]
-  pub xp: i64,
+  pub xp: f64,
 
   #[serde(rename = "XP/h", deserialize_with="de_from_str")]
-  pub xp_h: i64,
+  pub xp_h: f64,
 
 }
 
@@ -72,11 +72,11 @@ pub fn read_hunt_json(file: &Path) -> HuntInfo {
 }
 
 fn de_from_str<'de, D>(deserializer: D) -> 
-  Result<i64, D::Error> where D: Deserializer<'de> {
+  Result<f64, D::Error> where D: Deserializer<'de> {
 
   let s: String = Deserialize::deserialize(deserializer)?;
 
-  let f: i64 = s.replace(",", "").parse()
+  let f: f64 = s.replace(",", "").parse()
     .expect("Must be valid number");
   
   Ok(f)
